pyproject.toml
[project]
name = "eleguard"
version = "0.1.0"
description = "Raspberry Pi-based elephant detection and alert system for human-wildlife conflict prevention. "
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "inference-sdk>=0.9.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "picamera2>=0.3.12",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "roboflow>=1.1.3",
    "twilio>=8.5.0",
]
